+++ packages/charts/src/components/BarChart/BarRechart.stories.tsx
@@ -2,7 +2,7 @@ import { action } from '@storybook/addon-actions';
 import { boolean, text } from '@storybook/addon-knobs';
 import { BarChart } from '@ui5/webcomponents-react-charts/lib/next/BarChart';
 import React from 'react';
-import { complexDataSet, simpleDataSet } from '../../resources/DemoProps';
+import { complexDataSet, secondaryDimensionDataSet, simpleDataSet } from '../../resources/DemoProps';
 
 export default {
   title: 'Charts - Unstable /  BarChart',
@@ -83,6 +83,7 @@ export const defaultFormatterStory = () => (
     dataset={complexDataSet}
     dataLabelFormatter={(d) => `${d / 10}%`}
     xAxisFormatter={(el) => el / 10}
+    yAxisFormatter={(el) => el.slice(0, 3)}
     chartConfig={{
       xAxisUnit: '%',
       gridStroke: 'white',
@@ -131,4 +132,21 @@ export const withReferenceLineStory = () => (
 
 withReferenceLineStory.story = {
   name: 'With reference line'
+};
+
+export const withSecondardDimension = () => (
+  <BarChart
+    onDataPointClick={action('onDataPointClick')}
+    dataset={secondaryDimensionDataSet}
+    labelKey={'name'}
+    secondaryDimensionKey={'dimension'}
+    color={'lightblue'}
+    width={text('width', '95%')}
+    height={text('height', '70vh')}
+    chartConfig={{ dataLabel: true, barSize: 20 }}
+  />
+);
+
+withSecondardDimension.story = {
+  name: 'With secondary dimension'
 };