+++ packages/charts/src/components/RadarChart/RadarChart.tsx
@@ -81,7 +81,7 @@ const RadarChart: FC<RadarChartProps> = forwardRef((props: RadarChartProps, ref:
 
   const RadarDataLabel = useDataLabel(chartConfig.dataLabel, dataLabelCustomElement, dataLabelFormatter, false, false);
 
-  const marginChart = useChartMargin(dataset, labelKey, chartConfig.margin);
+  const marginChart = useChartMargin(dataset, yAxisFormatter, labelKey, chartConfig.margin);
 
   return (
     <ChartContainer