+++ packages/charts/src/resources/DemoProps.ts
@@ -123,3 +123,153 @@ export const simpleDataSet = [
     users: 280
   }
 ];
+
+export const secondaryDimensionDataSet = [
+  {
+    name: 'January / Month',
+    users: 100,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'February / Month',
+    users: 230,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'March / Month',
+    users: 240,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'April / Month',
+    users: 280,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'May / Month',
+    users: 100,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'June / Month',
+    users: 230,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'July / Month',
+    users: 20,
+    dimension: 'Harmnony Insight'
+  },
+  {
+    name: 'August / Month',
+    users: 220,
+    dimension: 'Harmnony Insight'
+  },
+  {
+    name: 'September / Month',
+    users: 200,
+    dimension: 'Harmnony Insight'
+  },
+  {
+    name: 'October / Month',
+    users: 250,
+    dimension: 'Intelligent Workplace'
+  },
+  {
+    name: 'November / Month',
+    users: 240,
+    dimension: 'Intelligent Workplace'
+  },
+  {
+    name: 'December / Month',
+    users: 280,
+    dimension: 'Pipeline Health Manager'
+  }
+];
+
+export const secondaryDimensionComposedDataSet = [
+  {
+    name: 'January',
+    users: 100,
+    sessions: 300,
+    volume: 756,
+    dimension: 'Harmnony Insight'
+  },
+  {
+    name: 'February',
+    users: 230,
+    sessions: 330,
+    volume: 880,
+    dimension: 'Harmnony Insight'
+  },
+  {
+    name: 'March',
+    users: 240,
+    sessions: 404,
+    volume: 700,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'April',
+    users: 280,
+    sessions: 80,
+    volume: 604,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'May',
+    users: 100,
+    sessions: 300,
+    volume: 756,
+    dimension: 'C4S Components'
+  },
+  {
+    name: 'June',
+    users: 230,
+    sessions: 330,
+    volume: 880,
+    dimension: 'Intelligent Workplace'
+  },
+  {
+    name: 'July',
+    users: 20,
+    sessions: 470,
+    volume: 450,
+    dimension: 'Intelligent Workplace'
+  },
+  {
+    name: 'August',
+    users: 220,
+    sessions: 180,
+    volume: 104,
+    dimension: 'Intelligent Workplace'
+  },
+  {
+    name: 'September',
+    users: 200,
+    sessions: 360,
+    volume: 879,
+    dimension: 'Intelligent Workplace'
+  },
+  {
+    name: 'October',
+    users: 250,
+    sessions: 500,
+    volume: 200,
+    dimension: 'Pipeline Health Manager'
+  },
+  {
+    name: 'November',
+    users: 240,
+    sessions: 404,
+    volume: 700,
+    dimension: 'Pipeline Health Manager'
+  },
+  {
+    name: 'December',
+    users: 280,
+    sessions: 80,
+    volume: 604,
+    dimension: 'Pipeline Health Manager'
+  }
+];