+++ packages/main/src/components/AnalyticalTable/hooks/useColumnsDependencies.ts
@@ -1,11 +1,10 @@
 export const useColumnsDependencies = (hooks) => {
   hooks.columnsDeps.push((deps, { instance: { state, webComponentsReactProperties } }) => {
     return [
+      ...deps,
       state.tableClientWidth,
       webComponentsReactProperties.scaleWidthMode,
-      webComponentsReactProperties.loading,
-      webComponentsReactProperties.selectionMode,
-      webComponentsReactProperties.noSelectionColumn
+      webComponentsReactProperties.loading
     ];
   });
 };